Create grid mouse hover trail
To enable each cell to change color when the mouse hovers on it, a CSS class called 'black' was created. in the JS file, under the makeCells() function (due to localization of the cells variable within that function), the class is added to each cell when the mouse goes over it.